dropped compatibility support of go-kit in discovery and log package
