+++ examples/skeleton/go.mod
@@ -6,4 +6,4 @@ toolchain go1.21.6
 
 replace github.com/cisco-open/go-lanai => ./../../../go-lanai
 
-require github.com/cisco-open/go-lanai v0.13.1-0.20240311221150-2525faca9482
+require github.com/cisco-open/go-lanai v0.13.1-0.20240324131552-56de4d9db8c1
